Editor data is getting store in redux store
