+++ src/common/constants.js
@@ -3,7 +3,7 @@ export default {
     'SESSION_SUCCESS': 'SESSION_SUCCESS',
     'ME_STATE_UPDATE': 'ME_STATE_UPDATE',
     //CONSTANTS FOR MESTORE
-    'UPDATE_LANGUAGE':'UPDATE_LANGUAGE',
+    'UPDATE_LANGUAGE': 'UPDATE_LANGUAGE',
     'REFRESH_TOKEN_SUCCESS': 'REFRESH_TOKEN_SUCCESS',
     'GET_FEATURE_VERSION': 'GET_FEATURE_VERSION',
     'GET_BOOT_SETTINGS': 'GET_BOOT_SETTINGS',
