+++ src/components/home/index.js
@@ -1,12 +1,12 @@
-import Editor from "../../editor/editor";
+import Editor from "../../editor/editorTS";
 import Sidebar from "../../sidebar/sidebar";
 
 function Home() {
   return (
-      <div className='memorymap-wrapper'>
-        <div className='sidebar'><Sidebar /></div>
-        <div className='editor'><Editor /></div>
-      </div>
+    <div className='memorymap-wrapper'>
+      <div className='sidebar'><Sidebar /></div>
+      <div className='editor'><Editor /></div>
+    </div>
   );
 }
